Retrieve prices asynchronously using cursor loader in item editing activity. Refactor purchase-related view into its own xml layout. Reuse prices and items xml layout. Use separate surfaces for purchase, item and prices in buying activity. Change text appearance and style in catalogue screen. Refactor codes in BuyingActivity. Change sql query for retrieving shopping List by joining Price table on price_id selection criteria.

+++ app/src/main/java/com/mirzairwan/shopping/domain/ToBuyItem.java
@@ -37,7 +37,7 @@ public class ToBuyItem implements Parcelable
 
     }
 
-    public ToBuyItem(int _idBuyItem, int quantity, Price selectedPrice, Item item, Date lastUpdatedOn)
+    public ToBuyItem(long _idBuyItem, int quantity, Price selectedPrice, Item item, Date lastUpdatedOn)
     {
         if(selectedPrice == null)
             throw new IllegalArgumentException("Price cannot be null");
